modify the path of codehub
