+++ pkg/microservice/aslan/core/workflow/handler/webhook.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/workflow/service/webhook"
 	internalhandler "github.com/koderover/zadig/pkg/shared/handler"
+	"github.com/koderover/zadig/pkg/tool/codehub"
 )
 
 // @Router /workflow/webhook [POST]
@@ -47,7 +48,7 @@ func ProcessWebHook(c *gin.Context) {
 		ctx.Err = processGithub(payload, c.Request, ctx.RequestID, ctx.Logger)
 	} else if gitlab.HookEventType(c.Request) != "" {
 		ctx.Err = webhook.ProcessGitlabHook(payload, c.Request, ctx.RequestID, ctx.Logger)
-	} else if webhook.HookEventType(c.Request) != "" {
+	} else if codehub.HookEventType(c.Request) != "" {
 		ctx.Err = webhook.ProcessCodehubHook(payload, c.Request, ctx.RequestID, ctx.Logger)
 	} else {
 		ctx.Err = webhook.ProcessGerritHook(payload, c.Request, ctx.RequestID, ctx.Logger)
